cgroup_integrate edited

+++ include/container-lib/cgroups.h
@@ -21,7 +21,7 @@ extern "C"{
 
 void echo_to_file(const char *path, const char *text, int len);
 int cgroup_verison(const char CGROUP_PATH[50]);
-void init_cgroup(long long MEM_SIZE, double TOTAL_CPU_PERCENTAGE,
+void init_cgroup(uint32_t MEM_SIZE, double TOTAL_CPU_PERCENTAGE,
                  const char CGROUP_ID[20], int CPU);
 void add_to_cgroup(pid_t pid, const char CGROUP_ID[20]);
 void deinit_cgroup(const char CGROUP_ID[20]);
